requirements.txt
python-telegram-bot==20.1
SQLAlchemy[asyncio]==2.0.4
asyncpg==0.27.0
jinja2==3.1.2
alembic==1.9.4
python-dotenv==0.21.1
pydantic==1.10.5

flake8==6.0.0
coverage==7.1.0
